pulumi/requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-azure-native>=2.0.0,<3.0.0
python-dotenv==1.0.0